added no client message to option 3 and 4

+++ Crap/main.py
@@ -37,14 +37,19 @@ what would you like to do?
                  for i in range(3):
                      client = upcoming[i]
                      print(f'{client["name"]} is due {client["report_due"]}') # print the list of clients and due dates
+             elif db.collection.count_documents({}) == 0:
+                 print('there are no clients')
              else:
-                for client in upcoming:
-                 print(f'{client["name"]} is due {client["report_due"]}') # print the list of clients and due dates
+                 for client in upcoming:
+                     print(f'{client["name"]} is due {client["report_due"]}') # print the list of clients and due dates
         elif int(user) == 4:
             select = input('Which client would you like to delete? ')
             query = {'name':f'{select}'}
-            db.collection.delete_one(query)
-            print(f'{select} has been deleted')
+            if db.collection.count_documents(query):
+                db.collection.delete_one(query)
+                print(f'{select} has been deleted')
+            else: 
+                print('that client does not exist')
 
 # exit contingency
         elif int(user) == 5:
